requirements.txt
smolagents
picsellia
Pillow
transformers
torch
numpy